apps/backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.1
django-cors-headers==4.5.0
scikit-learn==1.5.2
pyjwt[crypto]
cryptography
requests
numpy
openai
pdfplumber
python-docx
qdrant-client
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0
# Payment processing
stripe==7.12.0
# Social Media APIs
tweepy==4.14.0
python-linkedin-v2
# Local packages (editable installations)
-e ./packages/ai_core